+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Changed
+- Changed the `LpTim.wait` return type from `Result<(), Void>` to `Result<(), Infallible>`.
+
 ## [0.2.0] - 2021-11-11
 ### Added
 - Added two board support crates